+++ app/src/main/java/woowacourse/shopping/ui/product/ProductAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui
+package woowacourse.shopping.ui.product
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
