+++ app/src/main/java/woowacourse/shopping/ui/product/ProductViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui
+package woowacourse.shopping.ui.product
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
